fixing memory leaks, fixing warnings, adding iphone-related defines, replacing opengles 1 with 3 for iphone.

+++ trunk/core/smartbody/SmartBody/src/sk/sk_channel_array.h
@@ -102,7 +102,7 @@ public :
 //	void compress () { _channels.compress(); }
 
 	/*! Returns the number of channels in the array */
-	int size () const { return _channelList.size(); }
+	size_t size () const { return _channelList.size(); }
 
 	bool doesChannelExist(std::string name, SkChannel::Type t);
